+++ src/main/java/be/baur/sdt/statements/IfStatement.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import be.baur.sda.Node;
 import be.baur.sda.DataNode;
+import be.baur.sdt.StatementContext;
 import be.baur.sdt.TransformContext;
 import be.baur.sdt.TransformException;
 import be.baur.sdt.serialization.Statements;
@@ -26,8 +27,7 @@ public class IfStatement extends XPathStatement {
 	}
 
 
-	@Override
-	public void execute(TransformContext traco, StatementContext staco) throws TransformException {
+	@Override void execute(TransformContext traco, StatementContext staco) throws TransformException {
 		/*
 		 * Execution: create an XPath from the statement expression, set the variable
 		 * context and perform a Boolean evaluation. If the result is true, execute the