+++ src/main/java/be/baur/sdt/statements/OtherwiseStatement.java
@@ -2,6 +2,7 @@ package be.baur.sdt.statements;
 
 import be.baur.sda.Node;
 import be.baur.sda.DataNode;
+import be.baur.sdt.StatementContext;
 import be.baur.sdt.TransformContext;
 import be.baur.sdt.TransformException;
 import be.baur.sdt.serialization.Statements;
@@ -13,8 +14,7 @@ import be.baur.sdt.serialization.Statements;
  */
 public class OtherwiseStatement extends Statement {
 
-	@Override
-	public void execute(TransformContext traco, StatementContext staco) throws TransformException {
+	@Override void execute(TransformContext traco, StatementContext staco) throws TransformException {
 		/*
 		 * This method does nothing. Execution takes place in the context of the ChooseStatement.
 		 */