+++ src/main/java/demo.java
@@ -4,8 +4,8 @@ import java.io.PrintWriter;
 import be.baur.sda.SDA;
 import be.baur.sda.DataNode;
 import be.baur.sdt.SDT;
-import be.baur.sdt.Transform;
 import be.baur.sdt.TransformContext;
+import be.baur.sdt.statements.Transform;
 
 public class demo {