+++ src/test/java/test/TestSDTParser.java
@@ -7,10 +7,10 @@ import java.util.function.Function;
 
 import be.baur.sda.DataNode;
 import be.baur.sda.SDA;
-import be.baur.sdt.Transform;
 import be.baur.sdt.serialization.SDTParseException;
 import be.baur.sdt.serialization.SDTParser;
 import be.baur.sdt.statements.PrintStatement;
+import be.baur.sdt.statements.Transform;
 import be.baur.sdt.xpath.SDAXPath;
 
 public final class TestSDTParser {