+++ src/test/java/test/TestTransform.java
@@ -7,8 +7,8 @@ import java.io.Writer;
 import be.baur.sda.SDA;
 import be.baur.sda.DataNode;
 import be.baur.sdt.SDT;
-import be.baur.sdt.Transform;
 import be.baur.sdt.TransformContext;
+import be.baur.sdt.statements.Transform;
 
 public final class TestTransform {